Fixes
Inherits the following from core engine upgrade:

Upgrades bouncy castle and kestrel.
Fixes an issue where the status code property of the message info object is not initialized correctly.
Documents the fact that modifications to the message info class brings the burden of ensuring the validity of the object entirely on to the shoulders of the user.

Upgrades reference to WinDivertSharp, which now now longer uses pointer-to-reference conversion, in an attempt to resolve https://github.com/TechnikEmpire/WinDivertSharp/issues/3.